fix(main): transform request data
