+++ __tests__/index.spec.ts
@@ -1,8 +1,5 @@
 import axios, { AxiosInstance } from 'axios'
-import fs from 'fs'
-import path from 'path'
 import MockServer, { MockRoute, DataStore, Seeds, asyncResponse } from '~/src'
-import formToBuffer from './libs/formToBuffer'
 
 describe('initialize', () => {
   const mock = new MockServer()
@@ -109,40 +106,5 @@ describe('initialize', () => {
     const mockedData2 = await client.get(testPath)
     expect(mockedData2.data).toHaveLength(1)
     expect(mockedData2.data[0]).toEqual({ ...defaultValue, _id: expect.any(String) })
-  })
-
-  test.skip('post with multipart/form-data', async () => {
-    const testPath = '/test'
-    const name = 'sample-name'
-    const imageBinary = fs.readFileSync(path.join(__dirname, './assets/logo.png'))
-    const formData: [string, string | number | Buffer][] = [
-      ['name', name],
-      ['num', 123],
-      ['file', imageBinary]
-    ]
-    const imageDataURI = fs.readFileSync(path.join(__dirname, './assets/logo.b64'), 'utf-8')
-    const seeds: Seeds = { test: [] }
-    const route: MockRoute = [
-      {
-        path: testPath,
-        methods: {
-          get: () => asyncResponse(200, dataStore.getCollection('test').asyncFind({})),
-          post: ({ data: { file } }) =>
-            asyncResponse(201, dataStore.getCollection('test').asyncInsert({ file }))
-        }
-      }
-    ]
-
-    dataStore.init(seeds)
-    mock.setRoute(route)
-    const { boundary, buffer } = formToBuffer(formData)
-    const mockedData = await client.post(testPath, buffer, {
-      headers: { 'Content-Type': `multipart/form-data; boundary=${boundary}` }
-    })
-    expect(mockedData.data).toEqual({ file: imageDataURI, _id: expect.any(String) })
-
-    const mockedData2 = await client.get(testPath)
-    expect(mockedData2.data).toHaveLength(1)
-    expect(mockedData2.data[0]).toEqual({ file: imageDataURI, _id: expect.any(String) })
   })
 })
