+++ package.json
@@ -59,7 +59,6 @@
   },
   "dependencies": {
     "axios-mock-adapter": "^1.17.0",
-    "busboy": "^0.3.1",
     "chokidar": "^3.0.2",
     "file-type": "^12.2.0",
     "minimist": "^1.2.0",
